api/pyproject.toml
[tool.poetry]
name = "gdlk_api"
version = "0.1.0"
description = "GDLK HTTP API"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.1"
graphene-django = "^2.15.0"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
